requirements.txt
Cheetah==2.4.4
Django==1.6.1
Django-Select2==5.8.4
PAM==0.1.4
Pillow==3.1.1
PyYAML==3.10
argparse==1.2.1
chardet==2.1.1
colorama==0.2.5
configobj==4.7.2
constants==0.6.0
django-appconf==1.0.1
django-bootstrap3==7.0.1
django-debug-toolbar==1.0.1
django-filter==0.11.0
django-grappelli==2.5.1
django-imagekit==3.3
django-report-utils==0.3.15
django-tables2==1.0.7
djangorestframework==3.3.2
expressions==0.2.2
greenlet==0.4.2
gunicorn==17.5
html5lib==0.9999999
jsonpatch==1.3
jsonpointer==1.0
oauth==1.0.1
pilkit==1.1.13
prettytable==0.7.2
psycopg2==2.6.1
reportlab==3.3.0
requests==2.2.1
simplejson==3.8.2
six==1.10.0
ssh-import-id==3.21
urllib3==1.15.1
virtualenv==1.11.4
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5